+++ src/Contracts/ConditionCallableInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Tabuna\Map\Contracts;
+
+interface ConditionCallableInterface
+{
+    public function __invoke(mixed $value, object $source, ?object $target): bool;
+}