+++ src/Contracts/TransformCallableInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Tabuna\Map\Contracts;
+
+interface TransformCallableInterface
+{
+    public function __invoke(mixed $value, object $source, ?object $target): mixed;
+}