+++ tests/Dummy/MapAttrAdminTargetCondition.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tabuna\Map\Tests\Dummy;
+
+use Tabuna\Map\Contracts\ConditionCallableInterface;
+
+class MapAttrAdminTargetCondition implements ConditionCallableInterface
+{
+    public function __invoke(mixed $value, object $source, ?object $target): bool
+    {
+        return $target instanceof MapAttrAdminUserProfile;
+    }
+}