+++ tests/Dummy/MapAttrAdminUserProfile.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tabuna\Map\Tests\Dummy;
+
+class MapAttrAdminUserProfile
+{
+    public string $memberSince = '';
+
+    public ?string $ipAddress = null;
+}