+++ tests/Dummy/MapAttrApiPayload.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tabuna\Map\Tests\Dummy;
+
+class MapAttrApiPayload
+{
+    public string $product_name = '';
+
+    public string $price_amount = '';
+
+    public string $code = '';
+}