+++ tests/Dummy/MapAttrOnlineEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tabuna\Map\Tests\Dummy;
+
+class MapAttrOnlineEvent
+{
+    public string $title = '';
+}